feat: add confirmation dialogs for favorite removal and bulk removal
