+++ public/locale/en.json
@@ -645,6 +645,7 @@
   "are_you_still_watching": "Are you still watching?",
   "are_you_sure": "Are you sure?",
   "are_you_sure_cancel_planned_bed": "Are you sure you want to cancel the planned bed assignment?",
+  "are_you_sure_want_to_clear_favourite": "Are you sure you want to clear {{name}} from favourites?",
   "are_you_sure_clear_starred": "Are you sure you want to clear all starred?",
   "are_you_sure_mark_as_error_active_bed": "Are you sure you want to mark the active bed assignment as an error?",
   "are_you_sure_want_to_cancel_invoice": "Are you sure you want to mark this invoice as cancelled?",
